Bullets not explode when colliding with wall.  Added radomization to the animation that plays so that frequent effects like bullet fire doesn't look to repeditive.
Realized after it might be nice to add the randomization to the sprite render class itself.  That way this functionality doesn't need to be added to a bunch of different behaviours down the road.  It will also make extending this functionality with things like weighting easier.

It should maybe just detect when two animations have the same name and added them to a list.  When that animation is requested, it picks one from the list for you.