Increased the level from 20x20 to 200x200 and the frame rate dropped to 1fps...
Major bottle necks were rendering all the tiles in the level and doing collision checks against ever tile.

The fix the rendering bottleneck we now check that a tile is on screen before actually rendering it.

For collision checks we now quickly figure out which tiles to check against based on the fact that a tile's x,y position can be easily mapped to the 2d array of tiles.