Moved collision volume information to the Game Object level so that it can be used with GO that don't use the tile collision behaviour.
Thought about moving it to the SpriteRender behaviour since it will likely need to be tied to animation information at some point, but ultimately decided against it as it would then require GO to have a sprite to have collision which is not going to be true for things like trigger volumes.  

If animations need to change the volume based on animation they can just set it on a frame by frame basis in the Game Object class.