Main change was to add more functional support for drop shadows in the SpriteRender Behaviour.  It was implemented in such a way that it only worked for centered sprites.  This broke for this like the explosion that were rooted at the ground.  It now works in all cases.
However, the explosion animation exposed another limitation where the root was slightly recessed into the image, and so the shadow should be as well.  To handle those cases we now check for a special mAttachmentPoint called "Shadow".  If it finds that, it uses it instead of the motion root.

Also fixed a bug where the Kamikaze's Sprite definition had accidentally had its Ground attachment point removed.

Also speed up the explosion animation to make it look a little smoother.