Add new FiniteStateMachine Behaviour to handle to common flow book keeping a lot of behaviours may need.  This is an abstract class, so it is required that a Behaviour derive from this class to be used.
This is the case for the Civilian Behaviour which now derives from the FiniteStateMachine Behaviour rather than handling all that logic itself.

The second piece of the FiniteStateMachine is another abstract class, FSMState, which is how clients can implement logic within a state of the FiniteStateMachine.  Again an abstract class requiring clients to implement their own state logic.