Main purpose of this change was to fix Grenades so that they explode when hitting Enemies.  The Explosive Behaviour now checks to see if it is colliding with the types of objects it can damage and if so, Detonates.  It only does this if it is NOT set to manually explode, like in the case of the Kamikaze GameObject.
I also updated the Bullet GameObject to not use the Projectile Behaviour.  That Behaviour is made redundant by the additions to the Explosive Behaviour described above.

I also updated the player's Gun to have a shadow.  As a result, numerous issues were exposed with the current shadow positioning code, so that had to be fixed.  There is still some issues if scale is applied, but I will look into that later.

The scale issue was found because it turns out the Explosion Effect was scaling its sprite by %50.  I want that to be done offline, so now it is.

Fixed the safehouse positioning so that it lines up with the tile grid.

Played with Health and Damage on a number of Game Objects so that exploding Kamikaze could kill each other, but not instantly kill the player.