Stranded can how be rescued with Extraction Points.  Press L1 to drop a Becon, and all Stranded currently in the safe house will make a run for that point.  If they get there alive, they will be "resuced".
To support this I have added a new system for sending BehaviourMessages to all GameObject currently managed by the GameObjectManager.

On top of that, all BehaviourMessage objects now store the GameObject that sent the message.  In the case of an Extraction, this is used to determine the position that they should run to.