Created new Behaviour for Applying Damage on contact.  This is helpful for Explosions or other things like fire, where while it exists, if anything touches it, damage should be applied.  With this in place, the kamikaze behaviour no longer relies on objects touching the object to be hurt by the explosion.  Instead, the Explosion effect that it spawns can also do damage.
Fixed DamageFlash Behaviour so that GameObjects don't continue to flash after they have died.

Added debug functionality to manually slowing down updates.  By pressing left and right on the DPad, the user to force updates to be skipped.

To avoid being spammed with FPS debug display, the display has been changed to show in the update, rather than the renderer, that way it only shows once when the number of updates is reduced.

Motion Root data has been moved out of the SpriteRender Behaviour and into the base GameObject class.  This is because other things like the collision volume rely on that data. However, after doing this, I realized that the collision code assumed that the collision volume was in the center of the sprite, which is not true.  This has been fixed, and I have add a new offset "mCollisionOffset" which if defined, is used instead of mMotionRoot when placing collision volumes.

Added DPad support to the input manager.