Wall blocks are now induvidual GameObject. This is required so that we can add behaviours like Health to them.
Doing so added major performce issues do to the number of them being rendered. To fix, all GameObject must be on screen before they are rendered. To detect if they are on screen I have added a new Rectangle to every object, which represents its visible area on screen (like a bounding volume). There is no real logic associated with this; for now it just uses the CollisionRect scaled by 4.

With the Wall now a GameObject, a rare bug now became a lot more common (I think anyway). The issue was that seemingly at random civilians and enemies would suddenly die. It turns out that when a game object is recycled, it is repositioned, but its mCollisionRect is not repositioned until the end of the next frame. So objects would end up colliding with the old position of, say, an explision. This was very obvious with the wall because as bullets got recycled they would expode right away in the Player's gun because it thought it was still colliding with the wall.

To fix the issue I want to make sure mCollisionRect gets updated whenever mPosition changes. Unfortunately clients were setting the position directly (through the pOrientation property). I have removed the Orientation class all together and now clients must go through an interface to change an Object's position. When they do, mCollisionRect gets updated as well.

This change caused an issue with TileCollision behaviour as it relied on mCollisionRect being out of date to get a delta of the Player's movement. This is fixed by introducing a variable to GameObject which tracks the previous position of the object.

There is still a major performance issue, which appears to stem from collision checks against so many walls. Will likely need to change to some sort of tree or grid for static objects.