Walls now explode into a puff of smoke when they are destroyed.
Moved the health behaviour to Engine. It probably should have always been here, and now the new behaviour, RemoveTileOnDeath, can also live in Engine. Should audit existing Behaviours and move over any others that were in Game project due to dependancy on Health.

RemoveTileOnDeath is a simple Behaviour which just tells the Level to remove any collision at the position of a GameObject, when that GameObject hits zero health.

Finally removed the ".code" part in the GameObject name when loading from XML. It now matches the actual namespace.