Fixed bug with input being missed intermitently. It seems that the state of the GamePad can change mid update (although there is nothing saying so in the docs). When that happens there is a chance the Previous state we saved will be the same as the state in the next frame. As a result, some state changes will be missed (eg. going from released to press).
The solution is to simply store the GamePadState at the start of the frame, and use that cached value for the entire frame. That way, there is no change mid-update. That cached value becomes the previous value at the end of the frame, ensuring that no state changes are lost.