Added concept of GameObject Passes. These passes can be used to filter the updating and rendering of Behaviours. Using this, the ObjectPlacement behaviour only Renders when in the PLACEMENT pass, and only moves the cursor when in that pass.
Switching between the passes lives in the ObjectPlacement Behaviour so the entire Behaviour could not have its Update filter. Instead it just does a manual check inside its Update.

The functionality of moving a GameObject by its mDirection and mSpeed has been moved out of GameObject and into its own behaviour, SimpleMomentum. This is so that we can stop moving characters base on behaviour UpdatePasses.

https://app.asana.com/0/1194642625728/1260755855281