Static (stationary) and Dynamic (moving) objects are now tracked in seperate lists by the GameObjectmanager. The static list is actually a 2D array indexed by physical location in the world (cells). This was clients can quickly find a cell near a position in the world, and have a short list of all the surrounding objects. This improves performance greatly.
I am still noticing rare but significant drops in performance. I have done a little bit of profiling and reseting Timers seems to be pretty expensive, likely due to the fact that there are now thousands of them. Removed a redundant call to check if Timers are contained in a list before removing them, which seems to have helped a bit, but there are still performance issues to track down.

https://app.asana.com/0/1194642625728/2276964355414