Kamikaze B-Line.
Kamikaze now makes a proper B-Line for the player once he gets close. He will also do it if his path is blocked. The is a timer which forces him to explode after a few seconds. Eventually we may want him to explode on contact with walls, but the timer is so short it wouldn't be much different than what happens now.

Added a new animation for when the Kamikaze is in B-Line mode.

Discovered that the Kamikaze object's TileCollision behaviour was not functioning properly; he was walking right through walls when not using path finding. This was because the basic automated movement based on forward vector and speed was moved to its own behaviour, and that behavior was getting updated after the tile collision. As a result, the next update, the GameObjects current position and previous position were the same (since the last thing done by the GameObject in a frame is to take the current position and save it as the previous position), and the tile collision behaviour was confused since it thought the object had not moved. The fix is to simply put the tile collision behaviour before the movement behaviour.

StopWatch has been updated to use the Update Passes that GameObjects can use. This is so that timers will automatically stop when the game pauses. This was used for the Kamikaze countdownt to explosion.