Civilians are walking through walls when wandering. Also fixes Civilians often get stuck on edges of SafeHouse.
Both issues were caused by the fact that the Civilian has its center point right at the bottom of the image [4, 8]. When they Wander in the SafeHouse the are walking directly aligned with the tilemap. As a result, that center point (on the bottom edge of the Civilian) falls directly between 2 tiles. Our mapping from world space to tile ends up choosing the tile above the edge (and would choose the tile left of and edge on the X sides).

Civilians were wandering through walls because they actually thought they were in the wall due to the issue above, and as a result were finding open spots next to the wall. The visual though made it look like they were moving diagonally.

Civilians were getting stuck on the Edge of the SafeHouse because they thought they were 1 tile below the Safehouse, so the only valid move is to go straight up (not allow diagonal moves, and must move to a space inside the SafeHouse). When they arrive there, the issue described before said, again, that they were actually 1 tile below the SafeHouse, and the process repeats again and again.

To fix the issue I created a new method for check which Tile a GameObject is standing in. Unlike the only version, this takes more than just a position. It first tties to find a valid Tile based on position, but if the X or Y is between 2 tiles, it will instead use the CollisionRectangle center point which is usually more visually centered. If it is still on an edge it asserts and we will need to find a better solution.

https://app.asana.com/0/1194642625728/2975533245503
https://app.asana.com/0/1194642625728/2975533245499