MiniMap first pass implementation.
Basic MiniMap in place using proceedurally generated texture and just writing directly to pixel to mark objects on it. Right now is queries all the data it is interested in (eg. Player position) rather than actually being told which objects to mark. This will like change down the line. For now, it just marks the Player and the SafeHouse once it is placed.

I would also like to change it to resize based on map size so that it always takes the same amount of space on screen.

To help with this change, a new style of management has been added to GameObjectManager. There is now an additional list of GameObject which is indexed directly by Catagory. This allows us to quickly and easily do something like get all the Enemies in the world, or in the case of this change list, all th SafeHouse. It would be great to rework the GameObjectManager to use this for everything to make FindObjectsInRange() a lot faster.

https://app.asana.com/0/1194642625728/1260755855337