Bug: Major slowdown during game, especially in Debug.
Fix: A major contributer to the slowdown was the GameObjectPicker. Every frame it was doing an intersect test with every GameObject in the GameObjectManager. This was to print a little bit of data when hovering over an object. To start, this now only occurs when the debug display is active. And even then, we now only do the test when the user clicks, removing the hover-over-data functionality.

The entire Update of GameObjectPicker has been disable in Release builds.