MiniMap second pass implementation.
MiniMap is now told what GameObjects to mark, rather than it knowing internally what to mark. This works through a new message and a new Behaviour which can be added to GameObject: MarkOnMiniMap.

It no longer uses classifications at all, but that might be something interesting to add later.

In order to add Markers to the MiniMap in the opening frame of the game, I had to create a new virtual function for Behaviours: PostInitialization. This function gets called at the end of the Update where the GameObject was added to the GameObjectManager. This was you can depend on all GameObjects being in the GameObjectManager that were added on the same frame as you.

Also did some minor optimizations to GameObjectManager.Render().