Bug: Kamikaze are not releasing Nodes on death.
Fix: Just needed to clear the Destination of the PathFind Behaviour on death, but there was nothing to actually hook into at that point. To handle that, I added a new callback for when an object is removed from the GameObjectManager. I also renamed PostInitialization (called when an object is added to the GameObjectManager) to OnAdd.

https://app.asana.com/0/1194642625728/2994174578231