Additive sprite coloring.
Created functionality to tint sprites in a way that the color is added rather than the multiply version XNA ships. Since the default use of color is to multiply the texture RGB by the Color RBG, you can never add color; a black sprite will always be black.

This functionality required a custom shader; something not used on this game up to this point, so it required some changes to interfaces. The implementation of shader usage is quite barebones and really not ideal. We assume everything will use the same shader, and load that shader on the engine side. Really it should be customizable per sprite, and all fed in from the Engine, but that will be a bit of work (eg. SpriteRendererBehaviour is Engine side, so how does it know what fields to set).

The shader itself mimics the functionality of the default rendering in XNA and adds the tinting functionality on top. The old multiply style is still supported as well as a new tint style.

https://app.asana.com/0/1194642625728/2289094052657