Dynamically fill Safe House area.
Safe House can now be placed dynamically, filling up the structure the player is standing in. This is more complex than it initially sounds, as we want it to naturally be have as the user expects, meaning things like single tile openings get treated like doors between rooms, but don't count as a leak when they are an exit. I didn't want to force the user to place "door" tiles or something like that, so this seems to be the best option.

Note: This introduced a bug with the civilians where they now just stand in place when entering the Safe House. This is because they just don't want to leave the Safe House Area which is now broken into multiple tile sized collision volumes.

https://app.asana.com/0/1194642625728/3016475269892