Civilian find place to stand in safehouse rather than just wandering.
This exposed a bit of a bug in the Follow logic where in certain cases the civilians would get stuck on walls. This happened when the civilian position was slightly off, and so the path finder though he was standing on a wall and so the path search fails out right away. To fix, I use to collision center point instead, which is much closer to the center of the tile.