Basic path following now lives in its own Behaviour: PathFollow. Civilian has been updated to take advantage of this, and no longer does any of the path following logic itself.
To properly handle a Behaviour like this (one which should only run 'sometimes') I had to add the concept of a Behaviour being enabled/disabled. This is done on a per Behaviour Type basis. Right now, PathFollow states disabled, and when it is needed by the Civilian, it is turned on. When the Civilian wants to stop moving it is turned off.

To make things more readable, all Civilian state machines have been moved into their own files.

Unreleated, but Behaviour is now Abstract; it always should have been since there is no purpose in a having an empty Behaviour attached to a GameObject.

https://app.asana.com/0/1194642625728/3733006254384