Implemented proper BehaviourMessage naming convention. At its core, messages are named: <verb><noun>Message. For example:
SetHealthMessage
GetTileAtPositionMessage
ApplyDamageMessage

One slightly odd case is for messages that get sent when something happens, like taking dying. For those outgoing events we use the <verb> "On":

OnZeroHealthMessage
OnArrivedAtDestinationMessage

There has also been a change to the naming of members of BehaviourMessage objects. They are now post fixed with either _In for data that is passed into the OnMessage function, or _Out for data that gets filled in by the OnMessage function.

I have also changed the base BehaviourMessage class to be abstract and require all children to implement a Reset function.