Stranded find place to stand in safehouse: Avoid occupied tiles.
Civilian will now avoid occupied tiles when finding a place to stand in the SafeHouse. To support this functionality I changed the Tile a little bit. It now has a bitfield Attributes which will store things like "collison checked" and now "occupied"; stuff that used to be in TileType. TileType is now strickly tracking if a tile is solid or not.