Bug: Able to create safe house larger than max tiles.
Fix: The issue occured because the check to see if we have reached the max size of a single room did not account for how many tiles were used up prior to this room, so as long as this room was less than max size alone, it would succeed. It wouldn't propage too much because the check for whether a new room should be started did account for all previous tiles placed, so after the max was hit, the next room started would fail.

Room builder now accounts for tiles used on previous rooms when checking if the algorithm should continue.