Leveling System (WIP).
Added first stages of leveling system. User can stand on a Civilian who is in the SafeHouse, and a button hint is shown, instructing them to press X. If they do, that civilian will start researching a Health State Upgrade. After the required time has passed, the Player's max health will be increased.

Right now there is only one level of stat bonus, and it only works for health. The user does not choose a stat upgrade, and there is no pregress bar informing them that it is happening.

https://app.asana.com/0/1194642625728/1260755855308