Task: Job assignment for Stranded: Spawn scout through dialog box.
Added a popup which is opened when the user presses X while standing on a Civilian. The popup gives the Player the option of selecting HP level up, or convert to Scout (or back out with B).

Based on the selection made, the Civilian will perform the requested task.

I would still like to make the popup more generic so that it can be reused for all other Stranded types.

https://app.asana.com/0/1194642625728/1260755855310