Task: Scout Stranded type (full implementation): Send scouting through dialog box.
Added new template for StrandedPopup, "ScoutPopup". With it, the Player tells the Scout to go searching, rather than just the instant they press "X" as it was doing before.

Also fixed a minor issue with the way IconX was being placed. It was being offset by 1,1 to match the icons, but it didn't need to be. It was only needed because the Motion Root was incorrect.

https://app.asana.com/0/1194642625728/1260755855312