Task: Scout Stranded type (full implementation): Choose only valid targets.
Scout now checks with objects it attempt to scout, to see if they are "scoutable", before actually going to them. The button which starts the scouting will be disable in the case where there are no targets remaining.

Also made a minor change to the way text is positioned on the StrandedPopup so that it centers better with multi-line text.

https://app.asana.com/0/1194642625728/1260755855312