A* search is now almost completely generalized. Previously we relied on Level.Tile data for Graph information, but we want to be able to do A* searches on more than just Level.Tile objects, so that data needed to be abstracted.
All search logic is wrapped in a number of generic classes:

- GenericAStar.Planner
This is where the actual A* algorithm lives. It runs the algorithm on an abstract representation of the paths in the world, called a Graph.

- GenericAStar.Graph
Store all the connecting paths of the World. It stores those connections in a list of GraphNode objects.

- GenericAStar.GraphNode
Abstract Data Type which all GraphNode types will derive from. Stores all the general information about a single Node in a Graph as well as defining the interface which GenericAStar.Planner will use to perform A* with.

- GenericAStar.TileGraphNode
The sole implementation of GraphNode at the moment. This GraphNode wraps Level.Tile to mimic the functionality available prior to this change.

- GenericAStar.PathNode
Wrapper for GraphNode, but stores some additional data speific to Path Planning; previous Nodes, cost to reach this Node, etc.

This change also includes some really early work on an implementation of HPA* (which is what sparked the need to generalize our A* implementation in the first place).