WIP converting PathFind to use HPA*. Lots of changes made, but most significant are:
- GenericAStar.Planner now uses HPA* for path planning. This breaks things a little bit as some paths go through walls (it doesn't do low level Tile-by-Tile search when naviagating yet).
- NavMesh precomputes actual A* cost between all Neighbour GraphNode objects, so that a path an actual path cost can be quickly computed.
- New type of GraphNode, NavMeshTleGraphNode, which just adds some additonal debug information to TileGraphNode.
- Cluster is now an actual Graph.
- Links between GraphNode in NavMesh show cost by belending between 2 colors (white = cheap <-->  red = expensive).

There is a lot of unused functions added to NavMesh as originally I was thinking that I would need to temporarily add GraphNode objects to the Graph in order to perform searchs from arbitrary locations. That snowballed a bit, and I am planning to take a different approach. 

Instead of adding new Nodes the Graph, the Planner is going to do A* searches from the start and end positions to each GraphNode in their Cluster, without leaving the Cluster. The ones with the lowest cost will become the start and end of the Graph search. I'm pretty sure this should yeild decent results.

Alternatively, the start GraphNode could be dynamically created, and given 1-way links to all the other GraphNodes in the same Cluster, and then just do a regular search. The end is a little tougher because the GraphNodes in the Cluster need to know about it for the search to work.