Continued work on getting HPA* working. Things are mostly finished now. All units in the game use a combination of HPA* for long distance path finding, and tile base A* searches for pathing between GraphNodes in the HPA* path.
The one remaining piece is that the destination of the HPA* path search is just the existing GraphNode clostest to the destination position. This mostly works in a lot of cases, but the one that doesn't is Civilian moving to a standing point in the SafeHouse. Since the search is now using HPA*, it just finds the closest GraphNode, rather than the TileGraphNode at the given position.