Task: HPA* Path Finding > Search low level A* once high level path found.
PathPlanner Behaviour now supports searching at both high level NavMesh level and TileMap level at the "same time".

First it does the HPA*search to get the get the high level path to the destination. Once that is solved, it starts a TileMap level A* search to the first node in that high level path.

While this is happening, clients asking for the current best path will be given the high level path. Once the TileMap search finds the path to the first node of the high level path, that becomes the new best path.

From there is starts a search from that first node, to the second node at the TileMap level. However, while this search is happening the current "best" that clients see is not altered. Once it reaches that node, the current best is updated, and the process continues.

Task: HPA* Path Finding > Find exact destination (not just closest GraphNode).

This change also includes path finding using HPA* to the exact tile at the location requested. Prior to this it would just search for the closest GraphNode in NavMesh, but now it actually creates a temp GraphNode at the destination tile and that becomes part of the search graph. Once the search is no longer needed, the GraphNode is removed.

Since this GraphNode is removed after the search is not needed, the GraphNode always gets created, and existing GraphNode objects are never used. As a result, the debug checks to avoid GraphNode objects are the same location needed to be removed, since that can now come up in valid cases.

https://app.asana.com/0/1194642625728/4145631655092