Task: HPA* Path Finding > Path finding fails when placing tile on path.
When a tile is placed, the Cluster is cleared and recreated. However, the GraphNode that are temporarily placed as part of the PathFind behaviour are not part of the automated process of rebuilding a Cluster. As a result, after the tile is placed, the destination GraphNode no longer links to the other GraphNode in the Cluster and the path cannot be completed.

To get around this, GraphNode now track whether or not they are one of the temporary GraphNode. If they are, during the Clear process for a Cluster, they are not destroyed or removed (but all links to neighbours are). After that, the regular recreation process works to connect them with all the newly create GraphNode in the cluster.

https://app.asana.com/0/1194642625728/4145631655092