Task: Use Conditional Property instead of #if for things like Assert and Debug Draw.
Debug text functions are now all conditional to ALLOW_GARBAGE being defined. This means that any calls to it will not be compile into the program if that flag is not defined( eg. Release).

Created a new Property "ALLOW_DEBUG_SHAPES" to control whether or not the DebugShapeDisplay functionality gets compiled or not.

https://app.asana.com/0/1194642625728/4861757763058