Bug: Invalidating path causes search to restart from beginning.
When a path is invalidated we now reset the source position to the current position so that it does the search from there.

However, this exposed a bug in the path finding, in that if a path was invalidated it would actually start failing the search, but since it was often already solved, the GameObject would just continue to follow the path. This became more obvious on the cases where the GameObject was not constantly updating its Source and Destination (like FSMStatePatrol). It could easily be reproduced by placing a tile right in the destination.

To fix this, a couple behaviours now catch the path failing message and update the destination as needed.

https://app.asana.com/0/1194642625728/4915947913431